REVERT LATER
tryagain

+++ code/modules/mob/living/silicon/pai/death.dm
@@ -1,13 +1,12 @@
 /mob/living/silicon/pai/death(gibbed)
-	if(card)
-		card.removePersonality()
-		if(gibbed)
-			src.loc = get_turf(card)
-			qdel(card)
-		else
-			close_up()
-	if(mind)
-		qdel(mind)
-	..(gibbed)
+	if(stat == DEAD)
+		return
+	set_stat(DEAD)
+	mobility_flags = NONE
+	update_sight()
+	clear_fullscreens()
+
+	//New pAI's get a brand new mind to prevent meta stuff from their previous life. This new mind causes problems down the line if it's not deleted here.
+	remove_from_alive_mob_list()
 	ghostize()
 	qdel(src)
